src/requirements.txt
lorem
pyyaml
pyroute2==0.7.12